In Collection List--> Vector to do diiference operation find max

+++ 3- list/1-Vector/max_VectorApp.java
@@ -16,7 +16,7 @@ public class max_VectorApp{
 		while(i.hasNext()){
 			Object obj=i.next();
 			if((int)obj>max){
-				max=obj;
+				max=(int)obj;
 			}
 		}
